refactor(nika-runtime): split the seat receipt regression
Move the subscription-seat terminal receipt test behind its own module so
the runtime test root remains below the hard source-file line cap.
